requirements.txt
gradio~=3.6
torch~=1.12.1
torchvision~=0.13.1
pillow~=9.2.0